+++ main-server/src/http_server/routes/validator/json_schema/ndid.js
@@ -452,5 +452,21 @@ export default {
         required: ['name'],
       },
     },
+    '/ndid/add_suppressed_identity_modification_notification_node': {
+      body: {
+        properties: {
+          node_id: { type: 'string', minLength: 1 },
+        },
+        required: ['node_id'],
+      },
+    },
+    '/ndid/remove_suppressed_identity_modification_notification_node': {
+      body: {
+        properties: {
+          node_id: { type: 'string', minLength: 1 },
+        },
+        required: ['node_id'],
+      },
+    },
   },
 };